+++ changelog.md
@@ -145,3 +145,7 @@
 ## 24 Dec. 2021
 
 - Added Jump Field, Jump Restore Field, and Wall-Jump Block
+
+#### 24 Dec. 2021 #2
+
+- Fixed a bug where Force Block with addVel true doesn't work